Partial fix for ION-75: IonSystem.newDatagram() with imports doesn't actually use an import if it isn't in the catalog.
This corrects the problem for the old DOM, but the lite DOM encoding pathway is too convoluted for me to fix right now.  The main problem is that it encodes by piping a reader through a writer, thereby losing access to the symbol tables that are associated with the DOM.  It also seems to be wasting effort by recomputing local symtabs on the writer side, when the DOM already has that information and should just render it.

[GF:3853882:BrazilPerforce://brazil/src/shared/platform/IonJava/mainline]