ION-295: Improved symbol handling during IonSystemLite.iterate()
Iteration was calling populateSymbolValues on every top-level value, which constructs a fresh symbol table. No such table was ever promised, and that effort was taking 50% of CPU in some benchmarks of iteration.  Instead, we use the symtab that's maintained by the IonReader, which is exactly the symtab that was in fact used for the data.

https://cr.amazon.com/r/931596/

[GF:6245823:BrazilPerforce://brazil/src/shared/platform/IonJava/mainline]