Update Timestamp JavaDocs
- Every constructor is documented with what the precision of the resulting
Timestamp is, and what the definition of local offset parameter actually means
(it means minutes from UTC).
- Every static factory method is documented with what the precision of the
resulting Timestamp is, the Epoch time is defined in appropriate methods where
it is relevant.
- compareTo() and equals() methods have examples included, to inform clearly
what behavior is expected for equality and comparison checks.
- Timestamp.valueOf(CharSequence) is documented with the valid Ion format that
is expected, including a link to the W3C page and Ion wiki page.
- Timestamp.createFromUtcFields(...) is documented with examples of what
Timestamp returned when varying precisions of the same point in time is passed in.
- Added relevant TimestampTest test cases to ensure documented behaviors are
correct.

cr:jonker; https://cr.amazon.com/r/1481402/

[GF:7663064:BrazilPerforce://brazil/src/shared/platform/IonJava/mainline]