Add deprecation scripts into bin/
The deprecation scripts are intended to be used for parsing an audit log
to catch offending failed package builds that are still using deprecated
APIs. It will generate emails by storing ready-to-execute 'mail'
commands in a directory.

The mailer script is intended to be used for executing each of the
ready-to-execute 'mail' commands.

Please refer to the README for more information.

cr: jonker; https://cr.amazon.com/r/1557105/