Seperate shared symtab impl. out of UnifiedSymTab
Seperated shared symbol table implementation out of UnifiedSymbolTable
class. This is done so as to enable the most amount of parallelism.
The newly implemented SharedSymbolTable class is immutable, and hence
inherently thread-safe.

Note that the existing static factory methods in UnifiedSymbolTable
is not removed, but rather, all code-paths that call these methods
are re-routed to SharedSymbolTable's static factory methods. This is
because we want to iteratively refactor these classes, namely,
removing dead code is best left for another commit.

Fixed ION-364.
Tagged ION-379.

cr: jonker; https://cr.amazon.com/r/1888100/