Support for empty symbols, e.g. '' (#108)
Support for empty symbols, e.g. ''

Empty symbols are valid according to the spec so they need 
to be allowed. 
Removed usage of EmptySymbolException as it was only used in two cases:
* Empty symbols which are now supported
* Java null symbols which were converted to NPEs. 

EmptySymbolException was kept as deprecated to avoid breaking any 
client that depends on it 

issue: https://github.com/amzn/ion-java/issues/42