Implementing LST append through the DOM
The adapter now creates a read-only copy of mutable LST being imported.
The benefit of doing so is to keep the import structure in the DOM,
e.g. getImportedTables, and maintain separated the data regions that are
using the imported table and the derived table, example:

LST A
data region A
LST B extends A
data region A

when more symbols are added to data region A, after B is created, those
symbols should not be known to B.

Copying is not necessary for read only LSTs, as you can't add any more
symbols after derived tables are created, so they are referenced
directly as an optmization