Memleak gzip (#204)
* Fixing memleak when auto detecting gzipped bytes

Correctly closes the created IonReader in order to free the memory
allocated outside the heap by GzipInputStream.

For methods that return an iterator changed the signature to
return a CloseableIterator which is closeable permiting users to
close the iterator. This iterator also closes itself when there are
no more values next to return.

The gzip auto detect functionality should be removed from this package
as it is an orthogonal concern to reading/writing Ion data. Right now
our API has to close streams created by users to guaranteed that the
internal wrapped Gzip stream is correctly closed.

https://github.com/amzn/ion-java/issues/198