Makes the PooledBlockAllocatorProvider a singleton (#363)
Makes the PooledBlockAllocatorProvider a singleton

Many applications need to initialize IonWriters repeatedly for
short-lived tasks. For example, a webserver might create a new
IonWriter for each HTTP response that it needs to write. In such
cases, the bulk of the compute time needed to serialize the
response data is often spent initializing the byte array buffer
backing the IonWriter rather than performing the write logic itself.

Each time that a (e.g.) binary IonWriter is initialized, a new
PooledBlockAllocatorProvider (PBAP) is initialized with it to dole
out new byte array buffers. It attempts to reuse these byte arrays
wherever possible during its lifespan. Unfortunately, the PBAP only
lives as long as the IonWriter itself. If a new IonWriter is created,
a new PBAP (and therefore new byte arrays) are created as well.

This change modifies the PooledBlockAllocatorProvider (which was
already thread-safe) to be a globally shared singleton. This makes
repeatedly initializing new IonWriters substantially cheaper.