Compacts preallocated headers when possible (#401)
`IonRawBinaryWriter` can be configured to preallocate buffer bytes to
hold the length prefix of a new container or annotations wrapper. This
option trades processing speed for data size. Preallocating more bytes
makes the encoded data larger, but increases the chances that the
container's length (once calculated) can be written directly into the
existing buffer.

This patch modifies the writer such that, when finalizing the encoding
of a container or annotations wrapper, the writer will check to see if
the encoded length is small enough to fit in the type descriptor byte.
If it is, the encoded bytes are shifted backwards in the buffer to
eliminate the unnecessary preallocation. This means that the encoding of
small containers isn't penalized for having preallocation enabled.

Benchmarks showed no substantial difference in processing speed from the
previous release, so this data size improvement does not come at the
cost of throughput.