Fixes a bug in WriteBuffer#shiftBytesLeft (#416)
This patch fixes a bug in the `WriteBuffer` that caused the
`shiftBytesLeft` method to corrupt data if it was called such that the
final `Block` in the list was emptied AND the second-to-last block
reclaimed some capacity. This would only occur if the writer
was configured to use length preallocation greater than 0 and if they
were writing more than 32KB of Ion (the default block size) between
flushes.

This change accounts for that case and, if the last `Block`
is emptied, returns it to the pool for reuse.