Fix Annotated Dot in Sexp Parser Error (#440)
This fixes GH 434 by adding the proper transition to the state
machine.

There are two trivial code cleanups in related code. One drops
duplicate branch logic in the switch to the next state. The other
removes a boolean comparison by negating the variable instead.

I am adding a regression test to the ion-tests in a seperate PR.