Fix use of layout.buildDirectory
This is a DirectoryProperty, which in a few places was used or carried
forward as a String without being appropriately converted first.

That resulted in the use and creation of a project-root directory called
`gradle.api.file.Directory, fixed(class org.gradle.api.internal.file.DefaultFilePropertyFactory$FixedDirectory, `
which was then used by the `generateJarInfo` and `generateLicenseReport`
tasks.

This fixes it.